SCRUM-214 refactor: Auth domain 보완
