+++ src/main/java/Coordinate/coordikittyBE/domain/closet/service/ClosetService.java
@@ -3,9 +3,9 @@ package Coordinate.coordikittyBE.domain.closet.service;
 import Coordinate.coordikittyBE.domain.attach.repository.AttachRepository;
 import Coordinate.coordikittyBE.domain.auth.entity.User;
 import Coordinate.coordikittyBE.domain.auth.repository.UserRepository;
-import Coordinate.coordikittyBE.domain.closet.dto.ClosetCategorizationResponseDto;
-import Coordinate.coordikittyBE.domain.closet.dto.ClosetGetResponseDto;
-import Coordinate.coordikittyBE.domain.closet.dto.ClosetPostRequestDto;
+import Coordinate.coordikittyBE.domain.closet.dto.response.ClosetCategorizationResponseDto;
+import Coordinate.coordikittyBE.domain.closet.dto.response.ClosetGetResponseDto;
+import Coordinate.coordikittyBE.domain.closet.dto.request.ClosetPostRequestDto;
 import Coordinate.coordikittyBE.domain.closet.entity.Cloth;
 import Coordinate.coordikittyBE.domain.closet.repository.ClothDao;
 import Coordinate.coordikittyBE.domain.closet.repository.ClothRepository;
